Remove route from Segment class
Removed the route member from the Segment class.  Once upon a time, a
Segment was a series of Transitions, but it evolved into a slice of a
Route.  This made it impossible to use the segment timer without already
having a route defined.  This does not make that possible, but is a step
on the way.