Fix hanging when bytes come back in different chunk sizes than requested
Fixed a bug when reading from sd2snes via qusb2snes; the timer would
hang waiting for more data from qusb2snes even though it had already
been received.  This is (afaict) because qusb2snes was returning data in
64-byte chunks (as it received from sd2snes); when using qusb2snes with
retroarch, this does not happen, because retroarch sends responses back
in exactly the same sizes as in the request.