Improve recovery when there is packet loss
When a packet is dropped, the timer can get confused, because the next
packet read is for a previously requested memory address, not the memory
address that was just requested.

Mitigated this effect somewhat by reading all outstanding responses in
read_core_ram_multi before sending a request -- this ensures that no
memory address requested earlier can get confused with a response for
the new request.